requirements/base.txt
# Dependencias runtime/base (incluye paquetes de app y librerías de producción)
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==3.2.4
botocore==1.35.29
Brotli==1.2.0
certifi==2024.7.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==46.0.5
cssbeautifier==1.15.1
cssselect2==0.8.0
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
Django==4.2.27
django-appconf==1.1.0
django-cors-headers==4.5.0
django-crispy-forms==2.0
django-auditlog==3.0.0
django-import-export==3.2.0
django-multiselectfield==0.1.13
django-select2==8.4.1
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
docutils==0.19
EditorConfig==0.12.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fonttools==4.60.2
gunicorn==23.0.0
h11==0.16.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
idna==3.7
iniconfig==2.1.0
jmespath==1.0.1
jsbeautifier==1.15.1
json5==0.9.25
lxml==5.4.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.1.1
numpy==2.0.2
odfpy==1.4.1
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
pandas==2.3.1
pathspec==0.12.1
pillow==12.1.1
platformdirs==4.2.2
pluggy==1.5.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
PyMySQL==1.1.1
pypdf==6.8.0
pywebpush==2.1.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
regex==2023.12.25
reportlab==4.4.1
requests==2.32.4
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.4
svglib==1.5.1
tablib==3.6.1
tinycss2==1.5.0
tinyhtml5==2.0.0
tomli==2.0.1
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.6.3
uvicorn==0.34.0
weasyprint==68.0
webencodings==0.5.1
xhtml2pdf==0.2.17
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3.post1
num2words==0.5.12
python-json-logger==2.0.7
python-docx==0.8.11
htmldocx==0.0.6
docxtpl==0.16.7
drf-spectacular==0.27.0
markdown==3.6
sentry-sdk==2.39.0
